updated debian building system

+++ mkdeb.sh
@@ -1,5 +1,15 @@
+VERSION=$1
+
+if [ -z "$VERSION" ]
+then
+    echo "please set version"
+    exit 4
+fi
+
 cp equivs dist/build/puppetresources/ && \
+    cp ../language-puppet/ruby/calcerb.rb dist/build/puppetresources/ && \
     cd dist/build/puppetresources && \
+    sed -i -e "s/Version: .*/Version: $VERSION/" equivs && \
     equivs-build -f equivs && \
     mv puppetresources_*_amd64.deb ../../../ && \
     rm -f puppetresources*.changes && \
